+++ packages/social-metadata-forms/.babelrc
@@ -1,0 +1,23 @@
+{
+	"presets": [
+		[
+			"env",
+			{
+				"targets": {
+					"browsers": [ "extends @yoast/browserslist-config" ]
+				}
+			}
+		],
+		"react"
+	],
+	"plugins": [
+		"transform-object-rest-spread"
+	],
+	"env": {
+		"test": {
+			"plugins": [
+				"babel-plugin-styled-components"
+			]
+		}
+	}
+}
