+++ packages/social-metadata-forms/jest/setupTests.js
@@ -1,0 +1,6 @@
+import "raf/polyfill";
+import "jest-styled-components";
+import Enzyme from "enzyme";
+import EnzymeAdapter from "enzyme-adapter-react-16";
+
+Enzyme.configure( { adapter: new EnzymeAdapter() } );
