+++ packages/social-metadata-forms/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@yoast/social-metadata-forms",
+  "version": "1.0.0",
+  "main": "src/index.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Yoast/javascript.git",
+    "directory": "packages/social-metadata-forms"
+  },
+  "author": "Team Yoast",
+  "license": "GPL-3.0",
+  "private": false,
+  "scripts": {
+    "prepublishOnly": "rm -rf dist && cp -R src dist && cp package.json dist/package.json && json -I -f dist/package.json -e \"this.main='index.js'\" && cp .babelrc dist/.babelrc",
+    "test": "jest",
+    "lint": "eslint ."
+  },
+  "jest": {
+    "testURL": "http://localhost",
+    "testRegex": ".*Test.js$",
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+	],
+	"transform": {
+		"^.+\\.jsx?$": "babel-jest"
+	},
+	"transformIgnorePatterns": [
+		"/node_modules/(?!yoastseo|lodash-es).+\\.js$"
+	],
+	"setupTestFrameworkScriptFile": "<rootDir>/jest/setupTests.js"
+	},
+	"publishConfig": {
+	  "access": "public"
+	},
+  "devDependencies": {
+    "@yoast/browserslist-config": "^1.2.1",
+    "jest-styled-components": "^6.3.1",
+    "browserslist": "^4.7.3"
+  }
+}
