+++ packages/social-metadata-forms/src/index.js
@@ -1,0 +1,1 @@
+export { default as SocialMetadataPreviewForm } from "./SocialMetadataPreviewForm";
