+++ packages/social-metadata-forms/tests/redux/reducers/socialPreviewTest.js
