+++ packages/social-metadata-previews/src/index.js
@@ -1,3 +1,2 @@
 export { default as FacebookPreview } from "./facebook/FacebookPreview";
 export { default as TwitterPreview } from "./twitter/TwitterPreview";
-export { default as SocialMetadataPreviewForm } from "./shared/SocialMetadataPreviewForm";
